Move parseDom
